Former-commit-id: 24d7e450fd1fa08817fc0bf98ea6d7b5a7df6658

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerNumericIntegerQuestionCommand.cs
@@ -13,7 +13,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     {
         public int Answer { get; private set; }
 
-        public AnswerNumericIntegerQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, int answer)
+        public AnswerNumericIntegerQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, int answer)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.Answer = answer;
